Moving the location of the version_info class implementation. This keeps the location the same as before this feature addition which means import "just work", yet it separates the implementation into its own file where it really belongs.
